Refactor .map related UI
