+++ src/main/java/com/noticemedan/mappr/viewmodel/MapPaneController.java
@@ -1,17 +1,27 @@
 package com.noticemedan.mappr.viewmodel;
 
+import com.noticemedan.mappr.model.DomainFacade;
 import com.noticemedan.mappr.view.MapInfo;
+import com.noticemedan.mappr.view.util.FilePicker;
+import com.noticemedan.mappr.view.util.InfoBox;
+import io.vavr.control.Option;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
+import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ListView;
 import javafx.scene.layout.Pane;
 import javafx.scene.layout.StackPane;
+import javafx.stage.FileChooser;
+import javafx.stage.Stage;
 import lombok.Getter;
 import lombok.Setter;
+
+import javax.inject.Inject;
+import java.nio.file.Path;
 import java.util.ArrayList;
 
 public class MapPaneController {
@@ -21,12 +31,17 @@ public class MapPaneController {
 	@FXML Button mapPaneCloseButton;
 	@FXML Button loadMapButton;
 	@FXML Button deleteMapButton;
+	@FXML Button createMapButton;
 	@FXML StackPane noMapsYetPane;
 
 	@Setter
 	ObservableList<MapInfo> maps;
 	@Setter
 	MainViewController mainViewController;
+	DomainFacade domain;
+
+	@Inject
+	public MapPaneController(DomainFacade domainFacade) { this.domain = domainFacade; }
 
 	public void initialize() {
 		closeMapPane();
@@ -50,6 +65,7 @@ public class MapPaneController {
 
 		mapListView.getSelectionModel().selectedItemProperty().addListener(favoritePoiListener);
 
+		createMapButton.setOnAction(this::createMapFromOsm);
 		deleteMapButton.setOnAction(event -> {
 			maps.remove(mapListView.getSelectionModel().getSelectedItem());
 			if (maps.size() == 0) {
@@ -59,7 +75,6 @@ public class MapPaneController {
 		});
 
 		mapPaneCloseButton.setOnAction(event -> closeMapPane());
-
 	}
 
 	public void openMapPane() {
@@ -103,5 +118,19 @@ public class MapPaneController {
 	private void refresh() {
 		mapListView.setItems(maps);
 		if (maps.size() > 0) hideNoMapsYetPane();
+	}
+
+	private void createMapFromOsm(ActionEvent event) {
+		Stage stage = (Stage) this.mapPane.getScene().getWindow();
+		FilePicker picker = new FilePicker(new FileChooser
+				.ExtensionFilter("OSM Files (*.osm or *.osm.zip)", "*.osm", "*.osm.zip"));
+
+		Option<Path> path = picker.getPath(stage);
+		if (!path.isEmpty()) {
+			new InfoBox("Vi danner kortet i baggrunden - du vil få besked når det er færdigt.").show();
+			InfoBox onComplete = new InfoBox("Kortet er nu oprettet, og du har muligheden for at tilgå det fra menuen.");
+			InfoBox onFailed = new InfoBox("Der opsted en fejl under oprettelsen af kortet. Tilkald venligst dine nærmeste chimpanser.");
+			domain.buildMapFromOsmPath(path.get(), x -> onComplete.show(), x -> onFailed.show());
+		}
 	}
 }
