+++ src/main/resources/fxml/MapPane.fxml
@@ -8,18 +8,19 @@
 <VBox xmlns="http://javafx.com/javafx"
 	  xmlns:fx="http://javafx.com/fxml"
 	  fx:controller="com.noticemedan.mappr.viewmodel.MapPaneController"
-	  prefWidth="250.0"
+	  prefWidth="285.0"
 	  fx:id="mapPane">
 	<stylesheets>
 		<URL value="@../css/map-pane.css" />
 	</stylesheets>
-	<HBox prefHeight="55.0" prefWidth="200.0">
-		<Pane prefHeight="100.0" prefWidth="300.0">
+	<HBox prefHeight="55.0" prefWidth="240.0">
+		<Pane prefHeight="100.0" prefWidth="340.0">
 			<Label styleClass="paneTitle" text="Kort" />
-			<Button fx:id="mapPaneCloseButton" styleClass="closeButton" layoutX="215.0" layoutY="20.0" mnemonicParsing="false"/>
+			<Button fx:id="mapPaneCloseButton" styleClass="closeButton" layoutX="255.0" layoutY="20.0" mnemonicParsing="false"/>
 		</Pane>
 	</HBox>
 	<HBox fx:id="actionMenu" layoutY="55">
+		<Button fx:id="createMapButton" styleClass="actionMenuButton" text="Opret kort" />
 		<Button fx:id="loadMapButton" styleClass="actionMenuButton" text="Indlæs kort" />
 		<Button fx:id="deleteMapButton" styleClass="actionMenuButton" text="Slet kort" />
 	</HBox>
